Fix more SIZET_FORMAT bugs
With any luck this will now work on 32-bit Ubuntu.

+++ src/libcgraph/node.c++
@@ -90,8 +90,8 @@ void node::update(const mpz_class& value, ssize_t cycle)
 {
     if (cycle < _changed_cycle) {
         fprintf(stderr, "Last changed on %ld, now %ld\n",
-                _changed_cycle,
-                cycle
+                (long int)_changed_cycle,
+                (long int)cycle
             );
         abort();
     }
